[ll][dx12] merged data and state modules
